Reorganized configuration tests
Moved `MiniExcelOpenXmlConfigurationTests` and `MiniExcelAutoAdjustWidthTests` to own folder and separated models to own file
